v1.0.5: 体素算法重写(逐体素UV+超采样抗锯齿+double精度) + 浮点倍数 + 脚底对齐 + 速度持久化修复 + addmodel指引

+++ src/main/java/com/xxsx/builder/ai/ChatSession.java
@@ -17,7 +17,7 @@ public class ChatSession {
     private Runnable confirmAction;
     private CompletableFuture<?> pendingFuture;
     public String pendingBuildPath = null;
-    public int pendingBuildBaseScale = 150;
+    public float pendingBuildBaseScale = 150f;
     public int pendingBuildSpeed = -1; // -1=未指定, 用配置默认值
     private static final int MIN_KEEP_RECENT = 8;
